refactor: move mocks and stubs methods to a folder
