+++ .eslintrc.json
@@ -19,6 +19,7 @@
     "rules": {
         "no-useless-constructor": "off",
         "no-unused-vars": "off",
-        "@typescript-eslint/no-unused-vars": "off"
+        "@typescript-eslint/no-unused-vars": "off",
+        "@typescript-eslint/no-var-requires": "off"
     }
 }
