+++ jest.config.js
@@ -11,5 +11,6 @@ module.exports = {
   testEnvironment: 'node',
   transform: {
     '.+\\.ts$': 'ts-jest'
-  }
+  },
+  testMatch: ['**/?(*.)+(spec|test).ts']
 }
